pyproject.toml
[project]
name = "q24-hakernews-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.128.7",
    "groq>=1.0.0",
    "httpx>=0.28.1",
    "openai>=2.18.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
